Fix the format of the document names from an existing database

Fix the format of the document names from an existing database

fix_names(lang, full_names, language_df = NULL)

Arguments

lang
character defining the language, one of: "english", "portuguese", "spanish"

full_names
character, the list of full file.path to the document files

language_df
a named list of data.frame, at least one name should correspond to lang. The resulting data.frame should have a pdfs column

Value

a named list with two elements: names and full_names