Parse htlm page from Scopus to retrieve the abstract of a document

Parse htlm page from Scopus to retrieve the abstract of a document

get_scopusAbstract(urlScopus)

Arguments

urlScopus
URL of a document

Value

Text of the abstract of the document