Creates a weighted adjacency matrix of probability of citation between topics

Creates a weighted adjacency matrix of probability of citation between topics

make_topicNetwork(
  source_ids,
  citation_network,
  type = "NSF_specific",
  percentile.threshold = 0.9
)

Arguments

source_ids
source_ids

citation_network
an edge data.frame

Value

a weighted adjacency matrix of probability of citation between topics